Adding cli command for adding indexes
It's very basic for now, just accepting a name and a config and not doing any validation of them.
Set the basic cli framework in place, following how it's done in Lily, but using a different cli parsing lib: http://pholser.github.com/jopt-simple/
Before doing more work on the cli tools, we should now first finalize the model.